Use errors input instead of meta on serialize
If errors are passed in as an arugment to `serialize` they are not actually used in creating the `errors` section of the JSON:API output; instead `meta` is used. So, in order to actually get the errors to show up, you would need to do something along the lines of `json_api_doc.serialize(errors=errors, meta=errors)` but then the errors are in both the `errors` and `meta` section.

This commit fixes it so that if errors are passed in to a `serialize` call they are correctly added to the JSON:API output.